Move refund files to its folder and fix issues
